+++ Core/tests/Unit/Stubs/SampleGapicClass2.php
@@ -16,16 +16,20 @@
  * limitations under the License.
  */
 
-namespace Google\Cloud\Core\Tests\Unit\Helpers;
+namespace Google\Cloud\Core\Tests\Unit\Stubs;
 
 /**
  * This class is only intended to be used in tests to either mock
  * or execute some code in a predicted manner.
  */
-class SampleGapicClass1
+class SampleGapicClass2
 {
-    public function sampleMethod(&$counter)
+    public function sampleMethod(&$counter, $args)
     {
         $counter++;
+    }
+
+    public function sampleMethod2($arg1, $arg2, $optionalArgs = [])
+    {
     }
 }
